Alter file positioning
