--- JSON/genre.json
@@ -1,22 +1,0 @@
-[
-  {
-    "id": 25,
-    "name": "Ballet"
-  },
-  {
-    "id": 93,
-    "name": "Rock"
-  },
-  {
-    "id": 32,
-    "name": "Reguetton"
-  },
-  {
-    "id": 38,
-    "name": "Tapatio"
-  },
-  {
-    "id": 44,
-    "name": "Pop"
-  }
-]
