--- JSON/music_album.json
@@ -1,196 +1,0 @@
-[
-  {
-    "id": 243,
-    "archived": null,
-    "publish_date": "2021-11-16",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 19,
-        "name": "Ballet"
-      }
-    ]
-  },
-  {
-    "id": 738,
-    "archived": true,
-    "publish_date": "1996-09-13",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 19,
-        "name": "Ballet"
-      },
-      {
-        "id": 40,
-        "name": "Rock"
-      }
-    ]
-  },
-  {
-    "id": 738,
-    "archived": true,
-    "publish_date": "1996-09-13",
-    "on_spotify": "Y",
-    "genre": [
-
-    ]
-  },
-  {
-    "id": 765,
-    "archived": true,
-    "publish_date": "1111-11-11",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 40,
-        "name": "Rock"
-      }
-    ]
-  },
-  {
-    "id": 586,
-    "archived": true,
-    "publish_date": "1666-12-13",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 57,
-        "name": "Working"
-      }
-    ]
-  },
-  {
-    "id": 586,
-    "archived": true,
-    "publish_date": "1666-12-13",
-    "on_spotify": "Y",
-    "genre": [
-
-    ]
-  },
-  {
-    "id": 782,
-    "archived": true,
-    "publish_date": "1111-02-12",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 57,
-        "name": "Working"
-      },
-      {
-        "id": 32,
-        "name": "Erik"
-      }
-    ]
-  },
-  {
-    "id": 502,
-    "archived": true,
-    "publish_date": "1122-12-12",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 95,
-        "name": "Rock"
-      }
-    ]
-  },
-  {
-    "id": 252,
-    "archived": true,
-    "publish_date": "1111-01-01",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 25,
-        "name": "Ballet"
-      }
-    ]
-  },
-  {
-    "id": 802,
-    "archived": true,
-    "publish_date": "1212-12-12",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 93,
-        "name": "Rock"
-      }
-    ]
-  },
-  {
-    "id": 608,
-    "archived": true,
-    "publish_date": "1555-02-12",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 25,
-        "name": "Ballet"
-      }
-    ]
-  },
-  {
-    "id": 41,
-    "archived": true,
-    "publish_date": "1234-06-06",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 32,
-        "name": "Reguetton"
-      }
-    ]
-  },
-  {
-    "id": 41,
-    "archived": true,
-    "publish_date": "1234-06-06",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 32,
-        "name": "Reguetton"
-      }
-    ]
-  },
-  {
-    "id": 859,
-    "archived": true,
-    "publish_date": "1111-12-12",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 32,
-        "name": "Reguetton"
-      }
-    ]
-  },
-  {
-    "id": 384,
-    "archived": true,
-    "publish_date": "2000-12-12",
-    "on_spotify": "Y",
-    "genre": [
-      {
-        "id": 38,
-        "name": "Tapatio"
-      }
-    ]
-  },
-  {
-    "id": 704,
-    "archived": true,
-    "publish_date": "1111-11-11",
-    "on_spotify": "N",
-    "genre": [
-      {
-        "id": 44,
-        "name": "Pop"
-      }
-    ]
-  }
-]
