+++ app.rb
@@ -1,13 +1,12 @@
-# require_relative './game/json_data/game_module'
-require_relative './game/game_module'
-require_relative './module/music_module'
-require_relative './game/create'
-require_relative 'book'
-require_relative 'save_data'
+require_relative './modules/game_module'
+require_relative './modules/music_module'
+require_relative './modules/book_module'
+require_relative './preserve_data/save_data'
 
 class App
   include GameModule
   include MusicModule
+  include BookModule
 
   attr_reader :games, :authors
 
@@ -46,26 +45,18 @@ class App
     user_input = gets.chomp.to_i
 
     case user_input
-    when 1
-      list_books
-    when 2
-      music_list
-    when 3
-      list_games
-    when 4
-      show_genres
-    when 5
-      list_labels
-    when 6
-      list_authors
-    when 7
-      create_book
-    when 8
-      add_a_music
-    when 9
-      add_game
+    when 1 then list_books
+    when 2 then music_list
+    when 3 then list_games
+    when 4 then show_genres
+    when 5 then list_labels
+    when 6 then list_authors
+    when 7 then create_book
+    when 8 then add_a_music
+    when 9 then add_game
     when 10
       json_write
+      puts ''
       puts "Goodbye\n"
       exit
     else
@@ -73,50 +64,8 @@ class App
       select_input
     end
   end
-
-  def create_book
-    print 'Publisher : '
-    book_publisher = gets.chomp
-    print 'Cover State : '
-    book_cover_state = gets.chomp
-    print 'Publish Date(DD-MM-YYYY) : '
-    book_publish_date = gets.chomp
-    mybook = Book.new book_publisher, book_cover_state, book_publish_date
-
-    create(mybook)
-
-    @books << mybook
-
-    add_rel(mybook)
-
-    puts "Book saved.\n"
-
-    select_input
-    # save_book(book_publisher, book_cover_state, book_label_title, book_label_color, book_publish_date)
-  end
-
-  def list_books
-    puts ''
-    puts 'No book records.' if @books.empty?
-    @books.each do |book|
-      puts "[ Label: #{book.label.title} | Publisher: #{book.publisher} | Cover State: #{book.cover_state} ]"
-      # puts "ID: #{book.id}"
-      # puts "Publisher: #{book.publisher}"
-      # puts "Cover State: #{book.cover_state}"
-      # puts "Label: '#{book.label.title}', '#{book.label.color}'"
-      # puts "Publish Date: #{book.publish_date}"
-    end
-    select_input
-  end
-
-  def list_labels
-    puts ''
-    puts 'No label records.' if @labels.empty?
-    @labels.each { |label| puts "[ Title: #{label.title} | Color: #{label.color} ]" }
-    select_input
-  end
-  # rubocop:enable Metrics/CyclomaticComplexity
   # rubocop:enable Metrics/MethodLength
+  # rubocop:enable Metrics/CyclomaticComplexity
 
   def add_rel(item)
     @authors << item.author
