+++ item/book.rb
