+++ item/music_album.rb
@@ -1,4 +1,4 @@
-require_relative '../item'
+require_relative './item'
 require 'date'
 
 # Music Album class
