+++ item/rel/author.rb
@@ -1,5 +1,3 @@
-require_relative '../item'
-
 class Author
   attr_accessor :id, :first_name, :last_name
