+++ item/rel/label.rb
@@ -1,5 +1,3 @@
-require_relative 'item'
-
 class Label
   attr_accessor :id, :title, :color, :items
