+++ json_data/albums.json
@@ -1,3 +1,14 @@
 [
-
+  {
+    "id": 794,
+    "release": "2021-12-28",
+    "author_id": 140,
+    "fname": "sadfsd",
+    "lname": "fdsaf",
+    "label_id": 504,
+    "title": "fsdaf",
+    "color": "asdf",
+    "genre_id": 890,
+    "genre": "asf"
+  }
 ]
