+++ json_data/games.json
@@ -1,3 +1,14 @@
 [
-
+  {
+    "id": 261,
+    "release": "23-02-2021",
+    "author_id": 228,
+    "fname": "fdsgd",
+    "lname": "gdfsg",
+    "label_id": 776,
+    "title": "gdsfg",
+    "color": "gdfsg",
+    "genre_id": 734,
+    "genre": "gdsgf"
+  }
 ]
